Seperate the VM instance from the controls * creates a context that exposes an interface to control the js vm * vm controls now call context functions * vm provider manages redux actions and the vm
