+++ src/JavascriptVM/VMControls.tsx
@@ -1,0 +1,78 @@
+import { FunctionComponent, useContext } from "react";
+import React from "react";
+import { useSelector } from "react-redux";
+import { isExecuting } from "./vmSlice";
+import { getCode } from "./vmSlice";
+import { CommandBar, ICommandBarItemProps } from "@fluentui/react";
+import { VMContext, IVirtualMachine } from "./JavascriptVM";
+/**
+ * Renders a component that is responsible for controlling the VM according to the state
+ * of the application. Provides controls to conditionally start, stop, step, and run the VM.
+ */
+export const VMControls: FunctionComponent = () => {
+  const executing = useSelector(isExecuting);
+  const controls = useContext(VMContext) as IVirtualMachine;
+  const code = useSelector(getCode);
+
+  const startButton: ICommandBarItemProps = {
+    onClick() {
+      controls.start();
+    },
+    key: "start",
+    text: "Start",
+    iconProps: {
+      iconName: "Play",
+      className: "javascript-vm-controls--start-button",
+    },
+  };
+  const stepButton: ICommandBarItemProps = {
+    onClick() {
+      controls.step();
+    },
+    key: "next",
+    text: "Next",
+    iconProps: {
+      iconName: "Next",
+      className: "javascript-vm-controls--step-button",
+    },
+  };
+  const runButton: ICommandBarItemProps = {
+    onClick() {
+      controls.run();
+    },
+    key: "run",
+    text: "Run",
+    iconProps: {
+      iconName: "FastForward",
+      className: "javascript-vm-controls--run-button",
+    },
+  };
+  const stopButton: ICommandBarItemProps = {
+    onClick() {
+      controls.stop();
+    },
+    key: "stop",
+    text: "Stop",
+    iconProps: {
+      iconName: "Stop",
+      className: "javascript-vm-controls--stop-button",
+    },
+  };
+  const commandBarRunningItems: ICommandBarItemProps[] = [
+    stopButton,
+    stepButton,
+    runButton,
+  ];
+  const commandBarStoppedItems: ICommandBarItemProps[] = [
+    { ...startButton, disabled: !code },
+    { ...stepButton, disabled: true },
+    { ...runButton, disabled: true },
+  ];
+  return (
+    <div className="javascript-vm-controls">
+      <CommandBar
+        items={executing ? commandBarRunningItems : commandBarStoppedItems}
+      />
+    </div>
+  );
+};
